Extract the Box2D debug renderer to its own system

+++ core/src/dev/catcat/blitz/component/Quad.java
@@ -5,6 +5,6 @@ import com.artemis.Component;
 public class Quad extends Component {
     public String atlas;
     public String region;
-    public int w;
-    public int h;
+    public float w;
+    public float h;
 }
